Call swarm.process() synchronously
We need to call swarm.process() synchronously or we'll miss events, I
think.

The initial reason I moved this code into the callback for
`githubCurrentUser` was that I wanted the `username` in order to render
the highlights. But even if we get the `username` late, `virtual-dom`
should take care of updating the DOM.

cc @maxogden @mafintosh